191111 | django AUTH complete

+++ 웹 프로그래밍/django_form/articles/views.py
@@ -4,6 +4,7 @@ from .forms import ArticleForm, CommentForm
 from django.views.decorators.http import require_POST
 from IPython import embed
 from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 
 # Create your views here.
 def index(request):
@@ -11,6 +12,7 @@ def index(request):
     context = {'articles': articles}
     return render(request, 'articles/index.html', context)
 
+# 이런 식으로도 사용 가능하다! @login_required(login_url="/accounts/test/")
 @login_required
 def create(request):
     if request.method == "POST":
@@ -18,11 +20,13 @@ def create(request):
         form = ArticleForm(request.POST)
         # 폼이 유효한지 체크한다.
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
             # title = form.cleaned_data.get("title")
             # content = form.cleaned_data.get("content")
             # article = Article(title=title, content=content)
             # article.save()
+            article.user = request.user
+            article.save()
             return redirect("articles:detail", article.pk)
     else:
         form = ArticleForm()
@@ -37,43 +41,53 @@ def detail(request, article_pk):
     context = {'article': article, 'comment_form': comment_form, 'comments': comments}
     return render(request, 'articles/detail.html', context)
 
-# 이런 식으로도 사용 가능하다! @login_required(login_url="/accounts/test/")
-@login_required
 @require_POST
 def delete(request, article_pk):
-    article = get_object_or_404(Article, pk=article_pk)
-    article.delete()
-    return redirect("articles:index")
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article, pk=article_pk)
+        if request.user == article.user:
+            article.delete()
+        return redirect("articles:index")
+    return HttpResponse("You are Unauthorized", status=401)
 
 @login_required
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    if request.method == "POST":
-        form = ArticleForm(request.POST)
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            form.save()
-            # article.title = form.cleaned_data.get("title")
-            # article.content = form.cleaned_data.get("content")
-            # article.save()
-            return redirect("articles:detail", article_pk)
+    if request.user == article.user:
+        if request.method == "POST":
+            form = ArticleForm(request.POST)
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                form.save()
+                # article.title = form.cleaned_data.get("title")
+                # article.content = form.cleaned_data.get("content")
+                # article.save()
+                return redirect("articles:detail", article_pk)
+        else:
+            form = ArticleForm(instance=article)
     else:
-        form = ArticleForm(instance=article)
+        return redirect('articles:index')
     context = {'form': form}
     return render(request, 'articles/create.html', context)
 
 @require_POST
 def comments_create(request, article_pk):
-    article = get_object_or_404(Article, pk=article_pk)
-    comment_form = CommentForm(request.POST)
-    if comment_form.is_valid():
-        comment = comment_form.save(commit=False)
-        comment.article = article
-        comment.save()
-    return redirect("articles:detail", article_pk)
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article, pk=article_pk)
+        comment_form = CommentForm(request.POST)
+        if comment_form.is_valid():
+            comment = comment_form.save(commit=False)
+            comment.article = article
+            comment.user = request.user
+            comment.save()
+        return redirect("articles:detail", article_pk)
+    return HttpResponse("You are Unauthorized", status=401)
 
 @require_POST
 def comments_delete(request, article_pk, comment_pk):
-    comment = get_object_or_404(Comment, pk=comment_pk)
-    comment.delete()
-    return redirect("articles:detail", article_pk)
+    if request.user.is_authenticated:
+        comment = get_object_or_404(Comment, pk=comment_pk)
+        if request.user == comment.user:
+            comment.delete()
+        return redirect("articles:detail", article_pk)
+    return HttpResponse("You are Unauthorized", status=401)
